+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### Added
 - Extended error queue manager (`ServoErrorQueue`) to support standard drive errors (MOCO/COCO/SYSTEM).
 - Added standard error queue descriptors: `MOCO_ERROR_QUEUE`, `COCO_ERROR_QUEUE`, `SYSTEM_ERROR_QUEUE`.
-- Add `is_motion` and `custom_motion_rules` function in SafetyFunctions class.
+- Add `is_motion` and `follows_custom_motion_rules` in SafetyFunction class.
 
 ### Fixed
 - Clarify STO test logs.